+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 - Fixed setting the power roll characteristic if all applicable characteristics are negative.
 - Fixed missing story text from ability embeds. (#506)
 - Fixed kit speed bonuses not applying. (#511)
+- Added missing skills. (#518)
 
 ## 0.7.0 Foundry v13 Alpha